Fix ID command bug
As checking each ID used a status request to get the ID, the battery was being polled too frequently so would return a "battery busy" error, represented as -2. As the battery value is a uint_8t, this would underflow to 254.

Fix adds a backend method of requesting just the device ID.